que10.c

+++ C Programming /Assignments/Assignment 0/que10.c
@@ -1,12 +1,13 @@
 void main()
 {
-	int maths=80;
-	int science=85;
-	int history=90;
-	int english=97;
-	int marathi=95;
+	float maths=80;
+	float science=85;
+	float history=90;
+	float english=97;
+	float marathi=95;
 	float total=maths+science+history+english+marathi;
 	printf("total marks:%f",total);
 	float percent=(total/500)*100;
 	printf("percentage is:%f",percent);
+
 }
